[myfreecams] updated websocket script, added http-server and ignore case-sensitive usernames
the script will now use the http server
and not the websocket for the data of the channel.

it will only use the websocket to get the session id,
so we can use it for the http server.

case-sensitive usernames is not important anymore

For https://www.myfreecams.com/#UserNaME

works
streamlink https://www.myfreecams.com/#UserNaME best

will work now with this commit
streamlink https://www.myfreecams.com/#username best